chapter_001/src/main/java/ru/job4j/array/Turn.java
package ru.job4j.array;

/**
 * Turn
 * @author Denis Seleznev (d.selezneww@mail.ru)
 * @version $Id$
 * @since 0.1
 */
public class Turn {
    /**
     * back
     * @param array заданный массив.
     * @return перевернутый заданный массив.
     */
    public int[] back(int[] array) {
        for (int i = 0; i < array.length / 2; i++) {
            int tmp = array[i];
            array[i] = array[array.length - i - 1];
            array[array.length - i - 1] = tmp;
        }
        return array;
    }
}
